fix: alias was double deleted in parallel
it was fine in the tests, but it is depending on the order alias/index deletion happen finish, while running in parallel. In all cases , index deletion is deleting the alias, so we don't need to try to enforce it. This is a small breaking change on the ops output as wew remove 1 op